Standardize orders module structure
