+++ crates/model/src/orders/any.rs
@@ -19,7 +19,7 @@ use enum_dispatch::enum_dispatch;
 use serde::{Deserialize, Serialize};
 
 use super::{
-    base::Order, limit::LimitOrder, limit_if_touched::LimitIfTouchedOrder, market::MarketOrder,
+    Order, limit::LimitOrder, limit_if_touched::LimitIfTouchedOrder, market::MarketOrder,
     market_if_touched::MarketIfTouchedOrder, market_to_limit::MarketToLimitOrder,
     stop_limit::StopLimitOrder, stop_market::StopMarketOrder,
     trailing_stop_limit::TrailingStopLimitOrder, trailing_stop_market::TrailingStopMarketOrder,
