+++ crates/model/src/python/events/order/initialized.rs
@@ -32,7 +32,7 @@ use crate::{
     identifiers::{
         ClientOrderId, ExecAlgorithmId, InstrumentId, OrderListId, StrategyId, TraderId,
     },
-    orders::base::str_indexmap_to_ustr,
+    orders::str_indexmap_to_ustr,
     types::{Price, Quantity},
 };
