+++ crates/model/src/python/orders/limit_if_touched.rs
@@ -24,10 +24,7 @@ use crate::{
     identifiers::{
         ClientOrderId, ExecAlgorithmId, InstrumentId, OrderListId, StrategyId, TraderId,
     },
-    orders::{
-        LimitIfTouchedOrder,
-        base::{Order, str_indexmap_to_ustr},
-    },
+    orders::{LimitIfTouchedOrder, Order, str_indexmap_to_ustr},
     python::events::order::{order_event_to_pyobject, pyobject_to_order_event},
     types::{Price, Quantity},
 };
