+++ crates/model/src/python/orders/market_to_limit.rs
@@ -24,10 +24,7 @@ use crate::{
     identifiers::{
         ClientOrderId, ExecAlgorithmId, InstrumentId, OrderListId, StrategyId, TraderId,
     },
-    orders::{
-        MarketToLimitOrder,
-        base::{Order, str_indexmap_to_ustr},
-    },
+    orders::{MarketToLimitOrder, Order, str_indexmap_to_ustr},
     python::events::order::{order_event_to_pyobject, pyobject_to_order_event},
     types::Quantity,
 };
