+++ crates/model/src/python/orders/stop_limit.rs
@@ -27,10 +27,7 @@ use crate::{
     identifiers::{
         ClientOrderId, ExecAlgorithmId, InstrumentId, OrderListId, StrategyId, TraderId,
     },
-    orders::{
-        StopLimitOrder,
-        base::{Order, str_indexmap_to_ustr},
-    },
+    orders::{Order, StopLimitOrder, str_indexmap_to_ustr},
     python::{
         common::commissions_from_indexmap,
         events::order::{order_event_to_pyobject, pyobject_to_order_event},
