+++ crates/model/src/python/orders/trailing_stop_market.rs
@@ -25,10 +25,7 @@ use crate::{
     identifiers::{
         ClientOrderId, ExecAlgorithmId, InstrumentId, OrderListId, StrategyId, TraderId,
     },
-    orders::{
-        TrailingStopMarketOrder,
-        base::{Order, str_indexmap_to_ustr},
-    },
+    orders::{Order, TrailingStopMarketOrder, str_indexmap_to_ustr},
     python::events::order::{order_event_to_pyobject, pyobject_to_order_event},
     types::{Price, Quantity},
 };
